requirements.txt
netifaces==0.10.9
numpy==1.16.2
pigpio==1.44
Pillow==5.4.1
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
RPi.GPIO==0.7.0
spidev==3.4
qrcode==6.1
python-daemon==2.2.4

